gitdm: Report progress to stderr not stdout
When gitdm is used for generating text-only report with its output
redirected to a file, all is well aside from the clutter at the beginning
of that file -- a very long line with repeating "Grabbing changesets...".

Solve that by redirecting progress reporting to stderr. It also helps to
see the progress when you redirect gitdm output to a file.

Also, we don't have to flush stdout since stderr is unbuffered by default.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>
Signed-off-by: Jonathan Corbet <corbet@lwn.net>