Make tag matching stricter
If you commit a git changelog to your repository, gitdm will be confused by
all the added patch tags.  So make the patterns stricter to force them only
to match within the git log metadata - or so we hope.  There is still room
for confusion here; we really need to make grabpatch() smart enough to
split metadata and the diff.  Don't have time for that now.

This patch changes results slightly.  In the 2.6.36 cycle, there's a tag
reading:

    Original-Idea-and-Signed-off-by: Nicolas Pitre <nicolas.pitre@linaro.org>

Pre-patch gitdm would recognize that as a signoff; after the change it no
longer does.

Reported-by: Wolfgang Denk <wd@denx.de>
Signed-off-by: Jonathan Corbet <corbet@lwn.net>