Add the VirtualEmployer mechanism
A certain obnoxious developer wants his contributions to be split between
two employers.  So add the "VirtualEmployer" mechanism to make that
possible.  A virtual employer is defined with:

	VirtualEmployer ve-name
		nn% real-name
		...
	end

(This construct must appear in the main configuration file).  Developers
can be associated with the virtual employer in the usual way; at report
time, any changes credited to that employer will be split among the real
employers according to the percentages provided.

Signed-off-by: Jonathan Corbet <corbet@lwn.net>