Added option to get the stats from numstat instead of diff
The option --numstat of git log gives the statistics of
lines added and removed file.  Hence, it is not necessary
to parser a raw diff.

Another benefit, it is a less verbose log to be processed,
which helps to process long logs.  This also prepares the
code for counting the changes per file type.

Signed-off-by: Germán Póo-Caamaño <gpoo@gnome.org>