Filter out OpenStack Jenkins automated commits from the metrics
This change filters out the jenkins commits from all OpenStack git commit logs.  It does so at the point of processing within gitdm, rather than (for example) at the point of capture in the "do-it.sh" script.

This is somewhat brittle.  If the jenkins commit email address were to change, for example, this would stop filter.  Of course that would also be pretty obvious if we started seeing a bunch of Jenkins commits showing up again.

Change-Id: I2dc2b26778f3076561405ee3f0de58a8716e75cc
Reviewed-on: https://review.openstack.org/31666
Reviewed-by: Monty Taylor <mordred@inaugust.com>
Approved: Monty Taylor <mordred@inaugust.com>
Tested-by: Jenkins