Add header to output files and allow date ranges
This change adds a brief one line header to all default output files
to indicate when the analysis was generated.

It also allows the gitdm configuration file to use a date of the
format YYYY-MM-DD to specify the start or end date of a range of
commits, rather than only allowing a git changeset hash.  This is
valuable for analyzing projects which do not routinely or
consistently use tags, and when it is difficult to precisely
pinpoint the git commit that corresponds to a time-based release.

Change-Id: I899f6c6b046462addd3235e7797a3e1a63c8776c